Fix Hard Lockup in ktime_get_ts64 with debuggers
This patch fixes a hard system lockup that occurs when a system has been
held by a kernel debugger such as kgdb, kdb, or mdb on a multiprocessor
system with all processors suspended then released after 20-30 minutes
of debugging.

SUMMARY

An SMP multiprocessor capable kernel debugger must stop all processors
on a target system and trigger them to be held inside a NMI polling
routine while tracing and debugging active events on a target system.
This is typically accomplished by triggering an NMI interrupt via
an IPI call to each processor then directing the processors to enter
a suspend code loop from the NMI handlers.  While the processors
are held suspended, the system hardware timer will continue
to run as well as other hardware components.  Upon exiting
the debugger console, the processors are released and the system
is expected to resume normal operations.  In the case that causes this
bug, the Linux timekeeper subsystem sees a very large delta value
for the time in nanoseconds elapsed then attempts to process
this large number.  The bug is caused by a bit shift C instruction
for a variable defined as signed, which results in the compiler
using a SAR instruction rather than a SHR instruction that causes
the delta counter to wrap and backfill with sign bits.  The fix is
to instruct the compiler to output SHR instead of SAR to perform
the shift operation and to treat the operation as though the value
were unsigned rather than signed.

CONDITIONS THAT RESULT IN THIS LOCKUP

Holding the system processors suspended in the NMI handlers for
20 minutes to several hours while examining and debugging a target
system will trigger this bug upon exit from the debugger console
on either kgdb, kdb, or mdb.

Signed-off-by: Jeff Merkey <linux.mdb@gmail.com>